# autoassure-server-sdk/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "A2.Server | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/applications/{applicationId}/evidence-definitions": {
      "post": {
        "tags": [
          "EvidenceDefinitions"
        ],
        "operationId": "CreateEvidenceDefinition",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvidenceDefinitionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvidenceDefinitionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvidenceDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDefinitionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDefinitionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDefinitionResponse"
                }
              }
            }
          },
          "404": {
            "description": "No Application with the given applicationId exists in the\ncaller's Organization.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Returns 400 when the request fails a validation constraint."
          },
          "403": {
            "description": "Returns 403 when the caller's Organization is archived."
          }
        }
      },
      "get": {
        "tags": [
          "EvidenceDefinitions"
        ],
        "operationId": "ListEvidenceDefinitions",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvidenceDefinitionResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvidenceDefinitionResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvidenceDefinitionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/applications/{applicationId}/evidence-definitions/{evidenceDefinitionId}": {
      "patch": {
        "tags": [
          "EvidenceDefinitions"
        ],
        "operationId": "UpdateEvidenceDefinition",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "evidenceDefinitionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEvidenceDefinitionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEvidenceDefinitionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEvidenceDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDefinitionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDefinitionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDefinitionResponse"
                }
              }
            }
          },
          "404": {
            "description": "No EvidenceDefinition with the given evidenceDefinitionId exists in the\ncaller's Organization and Application.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Returns 400 when the request fails a validation constraint."
          },
          "403": {
            "description": "Returns 403 when the caller's Organization is archived."
          }
        }
      }
    },
...
    "/applications/{applicationId}/preconditions": {
      "post": {
        "tags": [
          "Preconditions"
        ],
        "operationId": "CreatePrecondition",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePreconditionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePreconditionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePreconditionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PreconditionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreconditionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreconditionResponse"
                }
              }
            }
          },
          "404": {
            "description": "No Application with the given applicationId exists in the\ncaller's Organization.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Returns 400 when the request fails a validation constraint."
          },
          "403": {
            "description": "Returns 403 when the caller's Organization is archived."
          }
        }
      },
      "get": {
        "tags": [
          "Preconditions"
        ],
        "operationId": "ListPreconditions",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreconditionResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreconditionResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreconditionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/applications/{applicationId}/preconditions/{preconditionId}": {
      "patch": {
        "tags": [
          "Preconditions"
        ],
        "operationId": "UpdatePrecondition",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "preconditionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreconditionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreconditionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreconditionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PreconditionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreconditionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreconditionResponse"
                }
              }
            }
          },
          "404": {
            "description": "No Precondition with the given preconditionId exists in the caller's\nOrganization and Application.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Returns 400 when the request fails a validation constraint."
          },
          "403": {
            "description": "Returns 403 when the caller's Organization is archived."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEvidenceDefinitionRequest": {
        "required": [
          "name",
          "description",
          "exampleValue"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "type": "string",
            "description": "Must not be empty or whitespace."
          },
          "description": {
            "maxLength": 500,
            "type": "string"
          },
          "exampleValue": {
            "maxLength": 10000,
            "type": "string"
          }
        },
        "description": "Request body to add an EvidenceDefinition to an Application's library."
      },
      "CreatePreconditionRequest": {
        "required": [
          "name",
          "valueSource",
          "exampleValue"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "type": "string",
            "description": "Must not be empty or whitespace."
          },
          "valueSource": {
            "$ref": "#/components/schemas/PreconditionValueSource"
          },
          "exampleValue": {
            "maxLength": 10000,
            "type": "string"
          }
        },
        "description": "Request body to add a Precondition to an Application's library."
      },
...
      "EvidenceDefinitionResponse": {
        "required": [
          "id",
          "name",
          "description",
          "exampleValue"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "exampleValue": {
            "type": "string"
          }
        },
        "description": "An EvidenceDefinition library item, as returned to the client."
      },
...
      "PreconditionResponse": {
        "required": [
          "id",
          "name",
          "valueSource",
          "exampleValue"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "valueSource": {
            "$ref": "#/components/schemas/PreconditionValueSource"
          },
          "exampleValue": {
            "type": "string"
          }
        },
        "description": "A Precondition library item, as returned to the client."
      },
      "PreconditionValueSource": {
        "type": "integer",
        "description": "Where a Precondition's value comes from at execution time."
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateEvidenceDefinitionRequest": {
        "required": [
          "name",
          "description",
          "exampleValue"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "type": "string",
            "description": "Must not be empty or whitespace."
          },
          "description": {
            "maxLength": 500,
            "type": "string"
          },
          "exampleValue": {
            "maxLength": 10000,
            "type": "string"
          }
        },
        "description": "Request body to edit an existing EvidenceDefinition."
      },
      "UpdatePreconditionRequest": {
        "required": [
          "name",
          "valueSource",
          "exampleValue"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "type": "string",
            "description": "Must not be empty or whitespace."
          },
          "valueSource": {
            "$ref": "#/components/schemas/PreconditionValueSource"
          },
          "exampleValue": {
            "maxLength": 10000,
            "type": "string"
          }
        },
        "description": "Request body to edit an existing Precondition."
      },
...
    }
  },
...
}